+++ src/main/resources/mapper/bookMapper.xml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="com.app.bookJeog.mapper.BookMapper">
-
-    <insert id="insertTempSelectedBooks">
+    <insert id="insertTempSelectedBook">
         insert into tbl_selected_book_temp (book_isbn)
-        values (#{bookisbn})
+        values (${isbn});
     </insert>
 </mapper>
