+++ src/main/resources/sql/book.sql
@@ -1,0 +1,2 @@
+insert into tbl_selected_book_temp (book_isbn)
+values (${bookisbn});
