+++ src/main/resources/static/js/admin/book_service.js
@@ -1,17 +1,23 @@
 const bookService= (() => {
-
-
     const getAllBook = async(callback) =>{
         let path =`/admin/books`;
         const response = await fetch(path);
         const bookList = await response.json();
-        console.log("📦 bookList 받음:", bookList);
         if(callback){
-            console.log("콜백 실행 확인")
             callback(bookList)
         }
     }
 
+    const tempSelectedBooks = async(chosenBookList, callback) => {
+        const params = new URLSearchParams();
+        chosenBookList.forEach(book => params.append("list", book));
 
-    return {getAllBook : getAllBook};
+        let path = `admin/temp-list?${params.toString()}`;
+        const response = await fetch(path);
+        const tempSelectedList = await response.json();
+        if(callback){
+            callback(tempSelectedList)
+        }
+    }
+    return {getAllBook : getAllBook, tempSelectedBooks:tempSelectedBooks };
 })();
